Restructuring Routes
Rewrote the the get_all_players and get_player routes to pull the data
from the database. Organized models into seperate folders. Created a
services folder to hold all the crud functions to simplify the logic in
the routers folder.
