+++ backend/app/scripts/populate_db.py
@@ -1,7 +1,6 @@
 import asyncio
 
 from app.db.database import Base, engine
-from app.db.models.models import *
 from app.scripts.core import main as insert_sports_and_leagues
 from app.scripts.nba.fetch_players import main as fetch_nba_players
 from app.scripts.nba.fetch_teams import main as fetch_nba_teams
